pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "foundry_agents"
version = "0.1.0"
description = "Foundry Agents SDK scaffolding"
readme = "README.md"
authors = [
  { name = "Your Name", email = "you@example.com" }
]
license = { file = "LICENSE" }
requires-python = ">=3.9"
keywords = ["azure", "ai", "sdk", "projects"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "python-dotenv",
  "azure-identity",
  "azure-ai-projects",
  "azure-ai-evaluation>=1.11.0",
  "promptflow-azure",
  "wikipedia",
  "openai"
]

[project.scripts]
foundry-agents = "foundry_agents.main:cli"

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.5.0",
  "mypy>=1.8.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["foundry_agents*"]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.9"
warn_unused_configs = true
warn_unused_ignores = true
strict = true
